requirements.txt
pygame
python-dotenv
openai-whisper
pyirsdk
sounddevice
wavio
numpy
groq
customtkinter
Pillow